chore: extract formatting utilities to lib/formatters.ts
